+++ src/RegRev/ExpressionContainer.php
@@ -10,12 +10,17 @@
 
 namespace RegRev;
 
-class ExpressionContainer implements \Iterator {
+/**
+ * Class ExpressionContainer
+ */
+class ExpressionContainer implements \Iterator
+{
     private $position = 0;
 
     private $expressions = array();
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->position = 0;
     }
 
@@ -24,23 +29,23 @@ class ExpressionContainer implements \Iterator {
         $this->expressions[] = $expression;
     }
 
-    function rewind() {
+    public function rewind() {
         $this->position = 0;
     }
 
-    function current() {
+    public function current() {
         return $this->expressions[$this->position];
     }
 
-    function key() {
+    public function key() {
         return $this->position;
     }
 
-    function next() {
+    public function next() {
         ++$this->position;
     }
 
-    function valid() {
+    public function valid() {
         return isset($this->expressions[$this->position]);
     }
 }