+++ src/RegRev/Metacharacter/Quantifier/ZeroOrMore.php
@@ -8,25 +8,21 @@
  * please view the LICENSE file that was distributed with this source code.
  */
 
-namespace RegRev\CharType;
+namespace RegRev\Metacharacter\Quantifier;
 
 /**
- * Class Number
- *
- * @package RevReg\Char
+ * Class ZeroOrMore,
+ * handles zero or more condition.
  */
-class NonAlnum extends CharType
+class ZeroOrMore extends OneOrMore
 {
     /**
      * {@inheritdoc}
      */
     public function generate()
     {
-        $characters = "!@£$%^&*()-_=+{}[];:'\"|<>?,./`~";
-        $randomIndex = rand(0, strlen($characters));
+        $quantity = $this->getQuanitity(0, 10);
 
-        return $characters[$randomIndex];
+        return $this->generateQuanitity($quantity);
     }
 }
-
-