+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # Changelog
 
-All Notable changes to RegexReverse will be documented in this file
+0.0.2
+==================
+
+  * Add more character types support
+  * Add non blank space support
+  * Add non alphanumeric support
+
+
+0.0.1 / 2015-04-10
+==================
+
+  * Initial release